vurderingTrygdeavgift.tsx var veldig mange linjer - deler den opp i mindre biter
