+++ src/felleskomponenter/stegvelger/stegKomponenter/ftrl/vurderingTrygdeavgift/index.ts
@@ -1,0 +1,3 @@
+import VurderingTrygdeavgift from "./vurderingTrygdeavgift";
+
+export default VurderingTrygdeavgift;
