+++ src/felleskomponenter/stegvelger/stegKomponenter/ftrl/vurderingTrygdeavgift/trygdeavgiftsgrunnlag/index.ts
@@ -1,0 +1,3 @@
+import Trygdeavgiftsgrunnlag from "./trygdeavgiftsgrunnlag";
+
+export default Trygdeavgiftsgrunnlag;
